requirements.txt
channels
Django
asgi_redis
redis
hiredis
python-dateutil
pytest-django